renamed legi.py package to dila2sql (#14)
