+++ cron/anomalies-cron.sh
@@ -3,9 +3,9 @@
 set -o pipefail
 
 cd "$(dirname "$0")/.."
-python -u -m legi.download ./tarballs
+python -u -m dila2sql.download ./tarballs
 echo "=> Starting importer..."
-python -u -m legi.importer legi.raw.sqlite tarballs --anomalies --anomalies-dir=anomalies --raw | tee -a legi.raw.log
+python -u -m dila2sql.importer legi.raw.sqlite tarballs --anomalies --anomalies-dir=anomalies --raw | tee -a legi.raw.log
 echo "=> Uploading anomaly logs..."
 rsync anomalies/ $1:~/anomalies/logs -rtv --chmod=F644
 echo "=> Generating index.html..."
