+++ cron/cron.sh
@@ -3,5 +3,5 @@
 set -o pipefail
 
 cd "$(dirname "$0")/.."
-python -m legi.download ./tarballs
-python -m legi.importer legi.sqlite ./tarballs | tee -a legi.log
+python -m dila2sql.download ./tarballs
+python -m dila2sql.importer legi.sqlite ./tarballs | tee -a legi.log
