+++ dila2sql/french.py
