+++ dila2sql/spelling.py
