+++ tox.ini
@@ -7,8 +7,8 @@ skipsdist=True
 [testenv]
 commands=
     pip install -q -r requirements.txt
-    python3 -m pytest {toxinidir}/tests --cov legi --cov-report=term-missing {posargs}
-    python3 -m pytest --doctest-modules legi
+    python3 -m pytest {toxinidir}/tests --cov dila2sql --cov-report=term-missing {posargs}
+    python3 -m pytest --doctest-modules dila2sql
     flake8 {toxinidir}
 deps=
     flake8
